I divided the bot logic into modules: logic, bot level, interface, messages from the bot, requests to the database. I add the nasldenik class to LobbyConsumer as an interface class with all the bot logic. To the 'bot_take_shot' and 'bot_create_new_game' methods I added the 'user' parameter with a User instance of the model. Refactored the src.game.consumers.mixins module (removed the 'BaseChooseWhoWillShotMixin' class and reworked the methods that depended on it). Wrote bot interfaces with separate and general logic. Made the 'determine_whoose_boards' method a separate synchronous function in the asynchronous shell. Classes for creating a lobby for playing with a bot and shooting a bot have been moved to the ..bots.bot_logic module. Moved the logic of messages from the bot to the ..bots.bot_message module.
