+++ src/game/bots/bot_interfaces.py
@@ -1,9 +1,13 @@
-from .bot_levels import LightBot, MediumBot, HighBot
+from . import bot_levels, bot_logic, bot_message
 
 
-class BotMessageInterface:
-    """"""
+class BotMessageInterface(bot_message.BotMainMessage):
+    """Interface of bot creades and sends message"""
 
 
-class BotDifficultyLevelsInterface(LightBot, MediumBot, HighBot):
+class BotDifficultyLevelsInterface(bot_levels.LightBot, bot_levels.MediumBot, bot_levels.HighBot):
     """Interface of bot difficulty levels for the game itself"""
+
+
+class GenericBotInterface(bot_message.BotMainMessage, bot_logic.BotTakeShot, bot_logic.BotCreatesNewGame):
+    """Interface class for all other bot mixins"""
